requirements_x.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
Cython==0.28.5
enum34==1.1.6
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.0
pandas==0.23.4
PlanOut==0.6.0
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
scikit-surprise==1.0.4
six==1.11.0
SQLAlchemy==1.2.10
SQLAlchemy-Utils==0.33.3
urllib3==1.23
websocket-client==0.48.0
Werkzeug==0.14.1
WTForms==2.2.1
